refactor : 리뷰 레파지토리 -> jpa,qsdl 공동 상속
