+++ Roy/umc9th/src/main/java/com/example/umc9th/domain/review/repository/ReviewQueryDsl.java
@@ -1,10 +1,10 @@
 package com.example.umc9th.domain.review.repository;
 
-import com.example.umc9th.domain.review.dto.ReviewResDto;
+import com.example.umc9th.domain.review.dto.ReviewResDTO;
 import com.querydsl.core.types.Predicate;
 
 import java.util.List;
 
 public interface ReviewQueryDsl {
-    List<ReviewResDto.ReviewItemDTO> searchReview(Predicate predicate);
+    List<ReviewResDTO.ReviewItemDTO> searchReview(Predicate predicate);
 }
