+++ Roy/umc9th/src/main/java/com/example/umc9th/domain/review/repository/ReviewQueryDslImpl.java
@@ -1,6 +1,6 @@
 package com.example.umc9th.domain.review.repository;
 
-import com.example.umc9th.domain.review.dto.ReviewResDto;
+import com.example.umc9th.domain.review.dto.ReviewResDTO;
 import com.example.umc9th.domain.review.entity.QReview;
 import com.example.umc9th.domain.store.entity.QStore;
 import com.querydsl.core.types.Predicate;
@@ -20,7 +20,7 @@ public class ReviewQueryDslImpl implements ReviewQueryDsl{
 
 
     @Override
-    public List<ReviewResDto.ReviewItemDTO> searchReview(
+    public List<ReviewResDTO.ReviewItemDTO> searchReview(
             Predicate predicate
     ){
         JPAQueryFactory queryFactory = new JPAQueryFactory(em);
@@ -30,7 +30,7 @@ public class ReviewQueryDslImpl implements ReviewQueryDsl{
 
         return queryFactory
                 .select(com.querydsl.core.types.Projections.constructor(
-                        ReviewResDto.ReviewItemDTO.class,
+                        ReviewResDTO.ReviewItemDTO.class,
                         review.id,
                         review.rate,
                         review.content,
