+++ Roy/umc9th/src/main/java/com/example/umc9th/domain/review/repository/ReviewRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Long>,ReviewQueryDsl {
 
     @Query("SELECT r FROM Review r " +
             "LEFT JOIN FETCH r.reviewPhotos " +
